#117 ♻️refactor: Project Header Sound Player
