+++ src/components/LandingPage/index.tsx
@@ -1,8 +1,8 @@
 import { memo, useEffect } from 'react';
 import LandingPageStyle from '@/layout/landingPage/LandingPageStyle';
 import { Cookies } from 'react-cookie';
-import LandingFooter from '../footer/LandingFooter';
-import LandingHeader from '../header/LandingHeader';
+import LandingFooter from '../footer/landingFooter';
+import LandingHeader from '../header/landingHeader';
 
 const LandingPage = memo(() => {
   const cookies = new Cookies();
