+++ src/components/createAvatar/index.tsx
@@ -3,8 +3,8 @@ import styled from 'styled-components';
 import Navbar from './navbar';
 import Voice from './voice';
 import Avatar from './avatar/index';
-import Header from '../header/ProjectHeader';
-import Footer from '../footer/ProjectFooter';
+import Header from '../header/projectHeader';
+import Footer from '../footer/projectFooter';
 import { useAppSelector, useAppDispatch } from '../../store/store';
 import Contents from './contents';
 import { useGetProjectDataQuery } from '@/api/useApi';
