+++ src/layout/projectContents/ProjectHeaderLayout.tsx
@@ -2,7 +2,7 @@ import React, { memo, useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
 import { useVideoSynthesisQuery, useAllListenQuery } from '@/api/useApi';
 import { useAppSelector, useAppDispatch } from '@/store/store';
-import SoundPlayer from '../SoundPlayer';
+import SoundPlayer from '../ProjectHeaderSoundPlayer';
 import styled from 'styled-components';
 import leftarrow from '/images/arrow-ios-left.png';
 import editpencil from '/images/edit-pencil.png';
